requirements.txt
torch
transformers
pillow
matplotlib
pytest